requirements.txt
flask
Flask-Cache
docker-py
uwsgi